long long->ll

+++ NumberTheory/mod pow.cpp
@@ -1,7 +1,7 @@
-long long mod_pow(long long a,long long b){
-  long long res=1;
+ll mod_pow(ll a,ll b){
+  ll res=1;
   while(b){
-    if(b&1){res=res*a%MOD;}
+    if(b&1){ res=res*a%MOD; }
     a=a*a%MOD;
     b>>=1;
   }
